Want simple performance tool
This tool allows us to pump IO full speed to supported storage URI's.
The intent of this is to be able to test performance without the need
for setting up any targets and, subsequently, benchmark tools.

Its no where near as sophisticated as fio, but thats also not the
intent. The intent is to get good numbers but also, to assist in
testing/developing the handling of IO errors and rebuild

+++ mayastor/src/core/mod.rs
@@ -63,7 +63,7 @@ pub enum CoreError {
     WriteDispatch {
         source: Errno,
         offset: u64,
-        len: usize,
+        len: u64,
     },
     #[snafu(display(
         "Failed to dispatch read at offset {} length {}",
@@ -73,7 +73,7 @@ pub enum CoreError {
     ReadDispatch {
         source: Errno,
         offset: u64,
-        len: usize,
+        len: u64,
     },
     #[snafu(display("Failed to dispatch reset",))]
     ResetDispatch {
@@ -87,12 +87,12 @@ pub enum CoreError {
     #[snafu(display("Write failed at offset {} length {}", offset, len))]
     WriteFailed {
         offset: u64,
-        len: usize,
+        len: u64,
     },
     #[snafu(display("Read failed at offset {} length {}", offset, len))]
     ReadFailed {
         offset: u64,
-        len: usize,
+        len: u64,
     },
     #[snafu(display("Reset failed"))]
     ResetFailed {},
